@@ -4,25 +4,20 @@
   "private": true,
   "homepage": "http://patrickrios.github.io/",
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.2",
-    "@fortawesome/fontawesome-free-brands": "^5.0.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-brands-svg-icons": "^5.15.2",
-    "@fortawesome/free-regular-svg-icons": "^5.15.2",
-    "@fortawesome/free-solid-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "emailjs-com": "^2.6.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "react-scroll": "^1.8.1"
+    "@emailjs/browser": "^4.3.3",
+    "@emailjs/nodejs": "^2.2.0",
+    "@types/react-scroll": "^1.8.10",
+    "framer-motion": "^11.0.8",
+    "html-react-parser": "^5.1.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.51.0",
+    "react-router-dom": "^6.22.2",
+    "react-scripts": "^5.0.1",
+    "react-scroll": "^1.9.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
@@ -45,7 +40,11 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.1.0",
-    "typescript": "^4.4.4"
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.24",
+    "@types/react": "^18.2.63",
+    "@types/react-dom": "^18.2.20",
+    "gh-pages": "^6.1.1",
+    "typescript": "^5.3.3"
   }
 }
